Added typescript and upgraded babel. Refactored all files to use typescript. Removed prototype overrides.
