+++ .babelrc
@@ -1,0 +1,9 @@
+{
+	"presets": [
+		"@babel/preset-typescript",
+		"@babel/preset-env"
+	],
+	"plugins": [
+		"@babel/plugin-proposal-class-properties"
+	]
+}
