+++ .eslintrc.yml
@@ -1,11 +1,17 @@
+root: true
 env:
   browser: true,
   node: true,
   shared-node-browser: true,
   es6: true
-extends: 'eslint:recommended'
+extends: 
+  'eslint:recommended'
+  'plugin:@typescript-eslint/eslint-recommended',
+  'plugin:@typescript-eslint/recommended'
+parser: '@typescript-eslint/parser'
 parserOptions:
   sourceType: module
+plugins: '@typescript-eslint'
 rules:
   accessor-pairs: error
   array-bracket-newline: 'off'
