+++ .gitignore
@@ -1,4 +1,5 @@
 *.log
+.vscode
 node_modules/
