+++ package.json
@@ -4,9 +4,9 @@
   "description": "Library for manipulating strings.",
   "main": "./dist/index.js",
   "scripts": {
-    "build": "babel ./src --presets babel-preset-es2015 --out-dir ./dist",
-    "test": "mocha",
-    "test:watch": "npm run test -- --watch",
+    "build": "npx babel ./src --out-dir ./dist",
+    "test": "npx mocha --require ./registerBabel.js ./test/**/*.ts",
+    "test:watch": "npm run test -- --watch --watch-extensions=js,ts",
     "prepublishOnly": "npm run build && npm run test"
   },
   "repository": {
@@ -20,29 +20,28 @@
     "url": "https://github.com/MyNameReallySux/string-utils/issues"
   },
   "homepage": "https://github.com/MyNameReallySux/string-utils#readme",
-  "babel": {
-    "presets": [
-      "es2015",
-      "stage-0"
-    ],
-    "plugins": []
-  },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
+    "@babel/cli": "^7.10.5",
+    "@babel/core": "^7.11.1",
+    "@babel/plugin-proposal-class-properties": "^7.10.4",
+    "@babel/preset-env": "^7.11.0",
+    "@babel/preset-typescript": "^7.10.4",
+    "@babel/register": "^7.10.5",
+    "@types/chai": "^4.2.12",
+    "@types/expect": "^24.3.0",
+    "@types/mocha": "^8.0.3",
+    "@typescript-eslint/eslint-plugin": "^3.9.1",
+    "@typescript-eslint/parser": "^3.9.1",
     "babel-preset-env": "^1.6.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-es2015-native-generators": "^6.6.0",
-    "babel-preset-stage-0": "^6.24.1",
     "chai": "^4.1.2",
-    "eslint": "^4.5.0",
+    "eslint": "^4.19.1",
     "eslint-loader": "^1.9.0",
-    "mocha": "^3.5.0",
-    "webpack": "^3.5.5"
+    "mocha": "^8.1.1",
+    "typescript": "^3.9.7",
+    "webpack": "^4.44.1"
   },
   "dependencies": {
-    "@beautiful-code/type-utils": "^0.1.3",
-    "babel-polyfill": "^6.26.0"
+    "@babel/polyfill": "^7.10.4",
+    "@beautiful-code/type-utils": "^0.1.3"
   }
 }
