+++ registerBabel.js
@@ -1,0 +1,1 @@
+require("@babel/register")({ extensions: ['.js', '.jsx', '.ts', '.tsx'] });
