+++ src/@types/@beautiful-code__type-utils.d.ts
@@ -1,0 +1,1 @@
+declare module '@beautiful-code/type-utils'
