+++ src/@types/global.d.ts
@@ -1,0 +1,3 @@
+declare type Either<A, B> = A | B
+declare type Optional<T, U = undefined> = Either<T, U>
+
