+++ src/index.ts
@@ -1,0 +1,2 @@
+export * from './StringUtils'
+export * from './StringStream'
