+++ test/mocha.opts
@@ -1,3 +1,3 @@
---recursive
+--require @babel/register
 --require ./test/test-header.js
---compilers js:babel-core/register
+--recursive
