+++ test/test-header.js
@@ -1,1 +1,1 @@
-import "babel-polyfill"
+import "@babel/polyfill"
